Long mode (64bit) Kernel

+++ src/PLlib/types.hpp
@@ -3,12 +3,12 @@
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
 typedef unsigned int uint32_t;
-typedef unsigned long uint64_t;
+typedef unsigned long long uint64_t;
 
 typedef char int8_t;
 typedef short int16_t;
 typedef int int32_t;
-typedef long int64_t;
+typedef long long int64_t;
 
 typedef unsigned int size_t;
-typedef unsigned int uintptr_t; // 32bit ptr
+typedef unsigned long long uintptr_t;
